Chore: mainly refactoring some frontend code
