+++ frontend/occupi-web/package.json
@@ -27,7 +27,8 @@
     "react-icons": "^5.2.1",
     "react-router-dom": "^6.23.1",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss": "^3.4.3"
+    "tailwindcss": "^3.4.3",
+    "vite-tsconfig-paths": "^4.3.2"
   },
   "devDependencies": {
     "@types/node": "^20.12.12",
