+++ frontend/occupi-web/src/Components/drawerComponent/DrawerComponent.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface DrawerComponentProps {
   isOpen: boolean;
   onClose: () => void;
