+++ frontend/occupi-web/src/Components/index.ts
@@ -1,0 +1,11 @@
+import DrawerComponent from "./drawerComponent/DrawerComponent";
+import GradientButton from "./gradientButtonComponent/gradientButton";
+import { Checkbox } from "./ui/checkbox";
+import InputBox from "./InputBox/InbutBox";
+
+export {
+    DrawerComponent,
+    GradientButton,
+    Checkbox,
+    InputBox
+}
