+++ frontend/occupi-web/src/assets/index.ts
@@ -1,0 +1,8 @@
+import loginpng from "./login.png";
+import loginsvg from "./login.svg";
+import occupi_logo from "./occupi-logo.svg";
+import otp from "./otp.png"
+
+export {
+    loginpng, loginsvg, occupi_logo, otp
+}
