+++ frontend/occupi-web/src/layouts/index.ts
@@ -1,0 +1,5 @@
+import Layout from './Layout'
+
+export {
+    Layout
+}
