+++ frontend/occupi-web/src/pages/index.ts
@@ -1,0 +1,8 @@
+import LoginForm from './Login/LoginForm'
+import OtpPage from './otp-page/OtpPage'
+import LandingPage from './landing-page/LandingPage'
+import Settings from './settings-page/Settings'
+
+export {
+    LoginForm, OtpPage, LandingPage, Settings
+}
