+++ frontend/occupi-web/vite.config.ts
@@ -1,11 +1,15 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import path from "path"
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    tsconfigPaths()
+  ],
 
 
   resolve: {
